Move polyfill preserver plugin to prod
